fix: result long 누적으로 오버플로우 방지

+++ src/main/java/calculator/interfaceImp/AddOperator.java
@@ -5,7 +5,7 @@ import calculator.calculatorInterface.Operator;
 public class AddOperator implements Operator {
     @Override
     public int operate(String[] tokens) {
-        int result = 0;
+        long result = 0;
         for (String t : tokens) {
             result += Integer.parseInt(t);
         }
@@ -14,6 +14,6 @@ public class AddOperator implements Operator {
             throw new IllegalArgumentException("정수 범위 초과");
         }
 
-        return result;
+        return (int) result;
     }
 }
